Add JwtClass, minor fixes.
