+++ OngProject/Core/Business/UsersBusiness.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
+using OngProject.Core.Helper;
 using OngProject.Core.Interfaces;
 using OngProject.Core.Models.DTOs;
 using OngProject.Entities;
@@ -43,7 +44,7 @@ namespace OngProject.Core.Business
 
         public async Task<Users> Insert(UserCreationDTO userDTO)  
         {
-            var pass = Encrypt(userDTO.Password);
+            var pass = JwtHelper.Encrypt(userDTO.Password);
             var user = new Users
             {
                 FirstName = userDTO.FirstName,
@@ -63,38 +64,9 @@ namespace OngProject.Core.Business
             throw new NotImplementedException();
         }
 
-        public static string Encrypt(string str)
-        {
-            SHA256 sha256 = SHA256Managed.Create();
-            ASCIIEncoding encoding = new ASCIIEncoding();
-            byte[] stream = null;
-            StringBuilder sb = new StringBuilder();
-            stream = sha256.ComputeHash(encoding.GetBytes(str));
-            for (int i = 0; i < stream.Length; i++) sb.AppendFormat("{0:x2}", stream[i]);
-            return sb.ToString();
-        }
+        
 
-        public string GetToken(Users user)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var key = Encoding.ASCII.GetBytes(_configuration["jwt:key"]);
-            var expiracion = DateTime.UtcNow.AddDays(1);
-            var tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity(
-                    new Claim[]
-                    {
-                        new Claim(ClaimTypes.NameIdentifier, user.Id.ToString()),
-                        new Claim(ClaimTypes.Email, user.Email),
-                        new Claim(ClaimTypes.Role, user.RolesId.ToString())
-                    }),
-                Expires = expiracion,
-                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
-            };
-
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            return tokenHandler.WriteToken(token);
-        }
+        
 
         
     }
